call sqrtf instead of sqrt, and atan2f instead of atan2 (float instead of double)

+++ src/types.c
@@ -15,7 +15,7 @@ vec3_t vec3_wrap_angle(vec3_t a) {
 }
 
 float vec3_angle(vec3_t a, vec3_t b) {
-	float magnitude = sqrt(
+	float magnitude = sqrtf(
 		(a.x * a.x + a.y * a.y + a.z * a.z) * 
 		(b.x * b.x + b.y * b.y + b.z * b.z)
 	);
